Fixed startup lag
Simple fix, great results. When trying to optimize for performance for this stuff try to scope to as few things as possible. So adding simple limits to things that are scoped to a lot can make it a lot faster because you scope to a less things overall. For example, here I limited the provinces so they only scope to them if they have the relevant pop, this reduces total scoping, making it faster. This is obvious but it's good to think about when making systems.