Mega Bugfixing (#282)
* Wealth window and trade svalues

* Governorship View!

* Trade Zones proof of concept & regional trade centers.

* WIP peace deal window

Very much not done, for some reason when I added the svalues they stopped working but im not too worried about it nothing is causing errors so it's fine.

* Fixed startup lag

Simple fix, great results. When trying to optimize for performance for this stuff try to scope to as few things as possible. So adding simple limits to things that are scoped to a lot can make it a lot faster because you scope to a less things overall. For example, here I limited the provinces so they only scope to them if they have the relevant pop, this reduces total scoping, making it faster. This is obvious but it's good to think about when making systems.

* Mega Bugfixing

90% of what I did here was simply purging almost all references to any vanilla things. Some things that are essential were spared though.

Co-authored-by: sobisonator <38443890+sobisonator@users.noreply.github.com>