Rewrote scripts in python, fixed stuff (#298)
* Wealth window and trade svalues

* Governorship View!

* Trade Zones proof of concept & regional trade centers.

* WIP peace deal window

Very much not done, for some reason when I added the svalues they stopped working but im not too worried about it nothing is causing errors so it's fine.

* Fixed startup lag

Simple fix, great results. When trying to optimize for performance for this stuff try to scope to as few things as possible. So adding simple limits to things that are scoped to a lot can make it a lot faster because you scope to a less things overall. For example, here I limited the provinces so they only scope to them if they have the relevant pop, this reduces total scoping, making it faster. This is obvious but it's good to think about when making systems.

* Mega Bugfixing

90% of what I did here was simply purging almost all references to any vanilla things. Some things that are essential were spared though.

* Fixed more bugs

* Found a solution to AI declaring wars. For now all ai have their aggressiveness reduced by 500,000. No need for a console command. I tested it, this works AI will not declare wars now but they still get into them from revolts.

* Adjusted the AI budget so they will now actually spend their money on buildings instead of just saving it all.

* Reduced all levy size modifiers significantly. Now base levy size is 1.5%. If it is any higher the game slows down because of massive army sizes.

* Fixed crash by readding the demand military access wargoal.

* Fixed a bunch of errors

* Movement Points Proof

* Also made it so AI doens't plan wars as much in defines. They don't need to take up performance planning since they will declare a different way.

* Industrial Goods gui work

* Trade Zones implemented

* Fixed 99% of gui issues and more trade zone work.

* Province View Changes

* Industrial View and added peace system files.

Also added the terrain gfx. Moscow and St Petersburg terrains need their gfx fixed.

* Finished industry buttons

* Movement Svalues up and running

* Rewrote scripts in python, fixed stuff

Co-authored-by: sobisonator <38443890+sobisonator@users.noreply.github.com>